While working on #8 I accidentally created an edge case where a task was killed, but it still showed as committed in the state of one of the pools. This was an easy mistake to make so I was thinking about ways to check for these kinds of mistakes in the future.

I came up with this script which creates a bunch of pools that are all connected to each other in different ways. Then I load them up and wait for them to finish and check the total count of processes in the VM and also check the state of all the pools to make sure we haven't left any processes alive or failed to cleanup our committed resources.