@mmmries 2022-02-14
I also ran the full suite of benchmarks on the current `main` branch and on this branch to check for any performance regressions that may have been caused during #7, #8 or this PR. But my run of the benchmarks showed no significant changes

# main branch

## Overhead (1x parallel)

```
Name                     ips        average  deviation         median         99th %
bare function     44667.51 K      0.0224 μs  ±8866.07%           0 μs      0.0900 μs
pooled              115.96 K        8.62 μs   ±136.71%        6.99 μs       37.99 μs
composed pools       78.74 K       12.70 μs    ±96.37%        9.99 μs       58.99 μs
handled               2.63 K      379.77 μs    ±29.33%      376.99 μs      665.99 μs
```

## Pool Throughput (16x parallel)

```
bare function        8185.04 K       0.122 μs  ±9362.36%           0 μs        0.99 μs
pooled function         7.77 K      128.66 μs    ±40.55%      121.99 μs      271.99 μs
composed pool           4.80 K      208.31 μs    ±38.15%      193.99 μs      408.99 μs
handled function        0.66 K     1510.36 μs    ±77.20%     1120.99 μs     5301.31 μs
```

# after all changes

## Overhead (1x parallel)

```
Name                     ips        average  deviation         median         99th %
bare function     40607.59 K      0.0246 μs  ±8012.42%           0 μs       0.100 μs
pooled              118.77 K        8.42 μs   ±126.65%           7 μs          21 μs
composed pools       79.33 K       12.61 μs    ±90.29%          10 μs          52 μs
handled               2.67 K      374.21 μs    ±28.07%         371 μs      582.49 μs
```

## Pool Throughput (16x parallel)

```
Name                       ips        average  deviation         median         99th %
bare function        9071.17 K       0.110 μs  ±8032.37%           0 μs           1 μs
pooled function         7.85 K      127.34 μs    ±42.80%         120 μs         280 μs
composed pool           4.80 K      208.53 μs    ±33.34%         194 μs         415 μs
handled function        0.68 K     1476.06 μs    ±76.12%        1102 μs        5131 μs
```