@mmmries 2022-02-27
I was also curious to try the new `OTP 25.0-rc1` so I re-ran the overhead benchmarks on that release candidate plus elixir 13.3 and I got these results:

```
Name                     ips        average  deviation         median         99th %
bare function     28986.12 K      0.0345 μs   ±594.95%           0 μs        1.00 μs
pooled              154.30 K        6.48 μs   ±118.24%        6.00 μs       10.00 μs
composed pools      111.85 K        8.94 μs   ±111.27%        8.00 μs       16.00 μs
handled               3.26 K      306.47 μs    ±25.91%      306.00 μs      450.00 μs
```

pooled alls went from `~120k => 154k` calls per second and composed pools went from `~80k => 111k` calls per second. Thanks BEAM team!